ensure ChangeProfilePicture not calls DeleteFile if file does not exists and SaveUserPicture throws

+++ src/domain/usecases/change-profile-picture.ts
@@ -31,7 +31,8 @@ export const setupChangeProfilePicture: Setup =
     try {
       await userProfileRepo.savePicture(userProfile);
     } catch {
-      await fileStorage.delete({ key });
+      if (file !== undefined) await fileStorage.delete({ key });
+      throw new Error();
     }
     return userProfile;
   };
